Move support and new_register outside of Spina
