+++ app/controllers/new_register_controller.rb
@@ -14,23 +14,23 @@ class NewRegisterController < ApplicationController
   end
 
   def suggest_register
-    @new_register = Spina::NewRegister.new
+    @new_register = NewRegister.new
   end
 
   def request_register
-    @new_register = Spina::NewRegister.new
+    @new_register = NewRegister.new
   end
 
   def thanks
   end
 
   def create
-    @new_register = Spina::NewRegister.new(new_register_params)
+    @new_register = NewRegister.new(new_register_params)
 
     if @new_register.valid?
       @zendeskService = ZendeskFeedback.new
       @response = @zendeskService.send_feedback(new_register_params)
-      redirect_to spina.new_register_thanks_path
+      redirect_to new_register_thanks_path
     else
       flash[:errors] = @new_register.errors
       if params[:new_register][:subject] == '[Request a Register]'
