--- app/models/spina/new_register.rb
@@ -1,9 +1,0 @@
-module Spina
-  class NewRegister
-    include ActiveModel::Model
-
-    attr_accessor :email, :name, :message, :subject
-
-    validates_presence_of :email, :name, :message
-  end
-end
